pyproject.toml
[tool.poetry]
name = "learn fastapi"
version = "0.1.0"
description = "'fastapi 학습하기'"
authors = ["Jin Yu/유진 <jin.yu@vuno.co>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
black = "^22.6.0"
flake8 = "^5.0.4"
python-jose = "^3.3.0"
bcrypt = "^3.2.2"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.9.2"}
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"
redis = "^4.3.4"
SQLAlchemy = "^1.4.40"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
